fix: int32→int64 for BSON property values, consistent array markers
Property values must be int64 in BSON — Studio Pro's type cache matches
by BSON type and crashes with "Sequence contains no matching element"
on int32. Array markers remain int32 (verified against Studio Pro BSON).

- Convert all property values from int32 to int64 across writer_*.go
- Add safeInt64() helper for settings serialization
- Remove debug logging left in writer_widgets_custom.go
- Fix 4 comments that incorrectly referenced int64 for array markers
- Add golden .mxunit files from Studio Pro for roundtrip testing

+++ sdk/mpr/writer_microflow_actions.go
@@ -373,7 +373,7 @@ func serializeMicroflowAction(action microflows.MicroflowAction) bson.D {
 			{Key: "$ID", Value: idToBsonBinary(string(a.ID))},
 			{Key: "$Type", Value: "Microflows$CloseFormAction"},
 			{Key: "ErrorHandlingType", Value: "Rollback"},
-			{Key: "NumberOfPagesToClose", Value: int32(a.NumberOfPages)},
+			{Key: "NumberOfPagesToClose", Value: int64(a.NumberOfPages)},
 		}
 		return doc
